requirements/common.txt
robotframework==4.0.3
robotframework-seleniumlibrary==5.1.3
